+++ src/network/utils/error-model.cc
@@ -250,7 +250,7 @@ RateErrorModel::DoCorruptByte (Ptr<Packet> p)
 {
   NS_LOG_FUNCTION_NOARGS ();
   // compute pkt error rate, assume uniformly distributed byte error
-  double per = 1 - std::pow (1.0 - m_rate, p->GetSize ());
+  double per = 1 - std::pow (1.0 - m_rate, static_cast<double> (p->GetSize ()));
   return (m_ranvar->GetValue () < per);
 }
 
@@ -259,7 +259,7 @@ RateErrorModel::DoCorruptBit (Ptr<Packet> p)
 {
   NS_LOG_FUNCTION_NOARGS ();
   // compute pkt error rate, assume uniformly distributed bit error
-  double per = 1 - std::pow (1.0 - m_rate, (8 * p->GetSize ()) );
+  double per = 1 - std::pow (1.0 - m_rate, static_cast<double> (8 * p->GetSize ()) );
   return (m_ranvar->GetValue () < per);
 }
